requirements.txt
Django==1.7.5
djangorestframework==3.3.2